Fiddling with the gc collector logic to get some good results.

+++ code/core/src/main/java/ikube/application/GCSnapshot.java
@@ -33,7 +33,7 @@ class GCSnapshot implements Serializable {
     /**
      * The amount of memory available after the collection, the maximum memory for the Jvm less the used portion.
      */
-    double available;
+    long available;
     /**
      * The ratio of used memory to maximum.
      */
